Added MakeWndClass in CWindow. Fixed working in AcceleratorTableGenerator.
git-svn-id: http://svn.purei.org/framework/trunk@171 6a6d099a-6a11-0410-877e-d5a07a98cbd2

+++ src/win32/AcceleratorTableGenerator.cpp
@@ -24,5 +24,5 @@ void CAcceleratorTableGenerator::Insert(unsigned int cmd, unsigned int key, unsi
 
 HACCEL CAcceleratorTableGenerator::Create()
 {
-    return CreateAcceleratorTable(&m_table[0], m_table.size());
+    return CreateAcceleratorTable(&m_table[0], static_cast<int>(m_table.size()));
 }
